Disable habitable room stats
We don't have any habitable room stats, so
we need to disable this data for now, or things
will break. I've tweaked the statistics partial
to set the `type` specifically, rather than
looping through an array with two objects, so
it's clearer about what we're doing, and what
we need to do when the habitable room data
is available to us.