Fix Node SSL issue
Assets cannot be compiled due to an issue caused by using @rails/webpacker which is no longer supported. Replacing it is a lot of work, so this is a short term fix.

The change in this commit affects local development. The equivalent fix in production is to add a Heroku config var:
`NODE_OPTIONS="--openssl-legacy-provider"`